Clean up godot::register module
