--- godot-core/src/builtin/meta/registration/mod.rs
@@ -1,11 +1,0 @@
-/*
- * Copyright (c) godot-rust; Bromeon and contributors.
- * This Source Code Form is subject to the terms of the Mozilla Public
- * License, v. 2.0. If a copy of the MPL was not distributed with this
- * file, You can obtain one at https://mozilla.org/MPL/2.0/.
- */
-
-//! Helpers used for registering various things with Godot.
-
-pub mod constant;
-pub mod method;
