+++ godot-core/src/lib.rs
@@ -5,40 +5,22 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
-mod registry;
-mod storage;
-
+// Note that a lot of those are public, but the godot crate still has the final say on what it wants to re-export.
+// Doing fine-grained visibility restrictions on every level is a useless maintenance chore.
 pub mod builder;
 pub mod builtin;
 pub mod classes;
 pub mod global;
 pub mod init;
 pub mod obj;
-
-#[deprecated = "Print macros have been moved to `godot::global`."]
-pub mod log {
-    pub use crate::global::{
-        godot_error, godot_print, godot_print_rich, godot_script_error, godot_warn,
-    };
-}
-
-#[doc(hidden)]
-#[path = "deprecated.rs"]
-pub mod __deprecated;
-#[doc(hidden)]
-pub mod private;
-
-pub use godot_ffi as sys;
-#[doc(hidden)]
-pub use godot_ffi::out;
-pub use registry::*;
-pub mod engine;
-
-/// Higher-level additions to the Godot engine API.
-///
-/// Contains functionality that extends existing Godot classes and functions, to make them more versatile
-/// or better integrated with Rust.
+pub mod registry;
 pub mod tools;
+
+mod storage;
+pub use godot_ffi as sys;
+
+// ----------------------------------------------------------------------------------------------------------------------------------------------
+// Generated code
 
 // Output of generated code. Mimics the file structure, symbols are re-exported.
 #[rustfmt::skip]
@@ -51,6 +33,9 @@ pub mod tools;
 mod gen {
     include!(concat!(env!("OUT_DIR"), "/mod.rs"));
 }
+
+// ----------------------------------------------------------------------------------------------------------------------------------------------
+// API version check
 
 macro_rules! generate_gdextension_api_version {
     (
@@ -89,3 +74,32 @@ generate_gdextension_api_version!(
         "4.1",
     },
 );
+
+// ----------------------------------------------------------------------------------------------------------------------------------------------
+// Hidden but accessible symbols
+
+/// Module which is used for deprecated warnings. It stays even if there is nothing currently deprecated.
+#[doc(hidden)]
+#[path = "deprecated.rs"]
+pub mod __deprecated;
+
+/// All internal machinery that is accessed by various gdext tools (e.g. proc macros).
+#[doc(hidden)]
+pub mod private;
+
+/// Re-export logging macro.
+#[doc(hidden)]
+pub use godot_ffi::out;
+
+// ----------------------------------------------------------------------------------------------------------------------------------------------
+// Deprecated modules
+
+#[deprecated = "Module has been split into `godot::classes`, `godot::global` and `godot::tools`."]
+pub mod engine;
+
+#[deprecated = "Print macros have been moved to `godot::global`."]
+pub mod log {
+    pub use crate::global::{
+        godot_error, godot_print, godot_print_rich, godot_script_error, godot_warn,
+    };
+}
