+++ godot-core/src/private.rs
@@ -9,7 +9,8 @@
 pub use crate::builtin::meta::trace;
 pub use crate::gen::classes::class_macros;
 pub use crate::obj::rtti::ObjectRtti;
-pub use crate::registry::{callbacks, ClassPlugin, ErasedRegisterFn, PluginItem};
+pub use crate::registry::callbacks;
+pub use crate::registry::plugin::{ClassPlugin, ErasedRegisterFn, PluginItem};
 pub use crate::storage::{as_storage, Storage};
 pub use sys::out;
