+++ godot/src/lib.rs
@@ -241,8 +241,14 @@ pub mod init {
 
 /// Register/export Rust symbols to Godot: classes, methods, enums...
 pub mod register {
-    pub use godot_core::property;
+    pub use godot_core::registry::property;
     pub use godot_macros::{godot_api, Export, GodotClass, GodotConvert, Var};
+
+    /// Re-exports used by proc-macro API.
+    #[doc(hidden)]
+    pub mod private {
+        pub use godot_core::registry::{constant, method};
+    }
 }
 
 /// Testing facilities (unstable).
