+++ itest/rust/src/register_tests/constant_test.rs
@@ -135,7 +135,7 @@ impl HasOtherConstants {
 impl godot::obj::cap::ImplementsGodotApi for HasOtherConstants {
     fn __register_methods() {}
     fn __register_constants() {
-        use ::godot::builtin::meta::registration::constant::*;
+        use ::godot::register::private::constant::*;
         // Try exporting an enum.
         ExportConstant::new(
             HasOtherConstants::class_name(),
